Refactor file transfer logic into separate module
